@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"io"
+	"log/slog"
 	"os"
 
 	"github.com/databricks/cli/libs/cmdio"
@@ -11,7 +12,6 @@ import (
 	"github.com/databricks/cli/libs/log"
 	"github.com/fatih/color"
 	"github.com/spf13/cobra"
-	"golang.org/x/exp/slog"
 )
 
 const (
@@ -52,11 +52,12 @@ func (l *friendlyHandler) coloredLevel(rec slog.Record) string {
 func (l *friendlyHandler) Handle(ctx context.Context, rec slog.Record) error {
 	t := fmt.Sprintf("%02d:%02d", rec.Time.Hour(), rec.Time.Minute())
 	attrs := ""
-	rec.Attrs(func(a slog.Attr) {
+	rec.Attrs(func(a slog.Attr) bool {
 		attrs += fmt.Sprintf(" %s%s%s",
 			color.CyanString(a.Key),
 			color.CyanString("="),
 			color.YellowString(a.Value.String()))
+		return true
 	})
 	msg := fmt.Sprintf("%s %s %s%s\n",
 		color.MagentaString(t),
@@ -76,16 +77,16 @@ type logFlags struct {
 func (f *logFlags) makeLogHandler(opts slog.HandlerOptions) (slog.Handler, error) {
 	switch f.output {
 	case flags.OutputJSON:
-		return opts.NewJSONHandler(f.file.Writer()), nil
+		return slog.NewJSONHandler(f.file.Writer(), &opts), nil
 	case flags.OutputText:
 		w := f.file.Writer()
 		if cmdio.IsTTY(w) {
 			return &friendlyHandler{
-				Handler: opts.NewTextHandler(w),
+				Handler: slog.NewTextHandler(w, &opts),
 				w:       w,
 			}, nil
 		}
-		return opts.NewTextHandler(w), nil
+		return slog.NewTextHandler(w, &opts), nil
 
 	default:
 		return nil, fmt.Errorf("invalid log output mode: %s", f.output)
