@@ -3,7 +3,7 @@ package log
 import (
 	"context"
 
-	"golang.org/x/exp/slog"
+	"log/slog"
 )
 
 type logger int
