@@ -1,10 +1,10 @@
 package log
 
 import (
+	"log/slog"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"golang.org/x/exp/slog"
 )
 
 func TestReplaceSourceAttrSourceKey(t *testing.T) {
